Lisasin templiidi, selle jaoks controlleri ja SecurtyConfigurationi.
